Moved to .netcore 2.0, moved from MSTest to NUnit
