--- csharp-ovh-test-suite/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="csharp-ovh" version="1.1.4" targetFramework="net451" />
-  <package id="ini-parser" version="2.2.4" targetFramework="net451" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net451" />
-</packages>
