--- csharp-ovh-test-suite/test-suite.playlist
@@ -1,1 +1,0 @@
-<Playlist Version="1.0"><Add Test="c:\users\lmarlin\documents\visual studio 2015\projects\csharp-ovh-test-suite\callstest.orderedtest" /><Add Test="csharp_ovh_test_suite.ClientWithtManualParams.ValidParams" /><Add Test="csharp_ovh_test_suite.ClientWithEnvironmentParams.ValidEndpoint" /><Add Test="csharp_ovh_test_suite.ClientWithEnvironmentParams.AllParams" /><Add Test="csharp_ovh_test_suite.ClientWithtManualParams.InvalidEndpointParamThrowsInvalidRegionException" /><Add Test="csharp_ovh_test_suite.ClientWithConfigFile.ValidConfigFileWithAllValues" /><Add Test="csharp_ovh_test_suite.ClientWithtManualParams.ValidEndpointParam" /><Add Test="csharp_ovh_test_suite.ClientWithConfigFile.ValidConfigFileWithEndpointOnly" /><Add Test="csharp_ovh_test_suite.ClientWithtManualParams.NoParamsThrowsConfigurationKeyMissingException" /><Add Test="csharp_ovh_test_suite.ClientWithConfigFile.InvalidConfigFile" /><Add Test="c:\users\lmarlin\documents\visual studio 2015\projects\csharp-clean\csharp-ovh-test-suite\callstest.orderedtest" /></Playlist>
