+++ csharp-ovh.sln
@@ -1,33 +1,48 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "csharp-ovh", "csharp-ovh\csharp-ovh.xproj", "{AA42BCD7-8E68-4F61-9C86-5C0B13E490FF}"
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "csharp-ovh", "csharp-ovh\csharp-ovh.csproj", "{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "csharp-ovh-test-suite", "csharp-ovh-test-suite\csharp-ovh-test-suite.csproj", "{E6849313-A85D-432F-8B93-0CB5352D63AD}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{50CF4A03-2630-483D-9F3A-F36D5B039A8D}"
-	ProjectSection(SolutionItems) = preProject
-		csharp-ovh\global.json = csharp-ovh\global.json
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "test", "test\test.csproj", "{050CFB9D-5A65-469C-9717-EB01CA2739B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AA42BCD7-8E68-4F61-9C86-5C0B13E490FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AA42BCD7-8E68-4F61-9C86-5C0B13E490FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AA42BCD7-8E68-4F61-9C86-5C0B13E490FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AA42BCD7-8E68-4F61-9C86-5C0B13E490FF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E6849313-A85D-432F-8B93-0CB5352D63AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E6849313-A85D-432F-8B93-0CB5352D63AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E6849313-A85D-432F-8B93-0CB5352D63AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E6849313-A85D-432F-8B93-0CB5352D63AD}.Release|Any CPU.Build.0 = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Debug|x64.ActiveCfg = Debug|x64
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Debug|x64.Build.0 = Debug|x64
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Debug|x86.ActiveCfg = Debug|x86
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Debug|x86.Build.0 = Debug|x86
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Release|x64.ActiveCfg = Release|x64
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Release|x64.Build.0 = Release|x64
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Release|x86.ActiveCfg = Release|x86
+		{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}.Release|x86.Build.0 = Release|x86
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Debug|x64.ActiveCfg = Debug|x64
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Debug|x64.Build.0 = Debug|x64
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Debug|x86.ActiveCfg = Debug|x86
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Debug|x86.Build.0 = Debug|x86
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Release|x64.ActiveCfg = Release|x64
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Release|x64.Build.0 = Release|x64
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Release|x86.ActiveCfg = Release|x86
+		{050CFB9D-5A65-469C-9717-EB01CA2739B8}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 EndGlobal
