--- csharp-ovh/Properties/AssemblyInfo.cs
@@ -1,22 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("chsarp_ovh")]
-[assembly: AssemblyDescription("Simple wrapper around OVH's Api mechanisms")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("OVH")]
-[assembly: AssemblyProduct("chsarp_ovh")]
-[assembly: AssemblyCopyright("Copyright ©  2016")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("aa42bcd7-8e68-4f61-9c86-5c0b13e490ff")]
