--- csharp-ovh/global.json
@@ -1,6 +1,0 @@
-﻿{
-  "sdk": {
-    "version": "1.0.0-rc1-update2"
-  },
-  "projects": ["csharp_ovh"]
-}
