--- csharp-ovh/project.json
@@ -1,18 +1,0 @@
-{
-  "version": "1.1.4-*",
-  "description": "chsarp-ovh Class Library",
-  "authors": [ "Luke Marlin" ],
-  "tags": [ "" ],
-  "projectUrl": "https://github.com/ovh/csharp-ovh",
-  "licenseUrl": "https://github.com/ovh/csharp-ovh/blob/master/LICENSE",
-  "frameworks": {
-    "net451": {
-      "dependencies": {
-        "ini-parser": "2.2.4"
-      }
-    }
-  },
-  "dependencies": {
-    "Newtonsoft.Json": "8.0.2"
-  }
-}
